Start cleaning up examples
