--- examples/RM2_WindIO.yaml
@@ -1,1 +1,0 @@
-nothing
