+++ examples/dev/SNL34m_InputsAeroDyn.yml
@@ -1,0 +1,49 @@
+
+general:
+  analysisType: unsteady # unsteady, steady, modal
+  turbineType: Darrieus #Darrieus, H-VAWT, ARCUS
+
+designParameters:
+  eta: 0.5 # blade mount point ratio, 0.5 is the blade half chord is perpendicular with the axis of rotation, 0.25 is the quarter chord, etc
+  Nbld: 2 # number of blades
+  Blade_Radius: 17.1 # blade radius m
+  Blade_Height: 41.9 # blade height m
+  towerHeight: 0.5 # m tower extension height below blades
+
+operationParameters:
+  rho: 0.94 # air density
+  Vinf: 10.1 # m/s #optional, supersceeded if ifw=true
+
+turbulentInflow:
+  ifw: false
+  WindType: 3
+  windINPfilename: /data/turbsim/115mx115m_30x30_20.0msETM.bts
+  ifw_libfile: nothing
+
+controlParameters:
+  controlStrategy: constantRPM # TODO: incorporate the others
+  RPM: 34.0 #RPM
+  numTS: 100 #
+  delta_t: 0.05 # s
+
+
+AeroParameters:
+  Nslices: 35 # number of OWENSAero discritizations #TODO: AD parameters
+  ntheta: 30 # number of OWENSAero azimuthal discretizations
+  AModel: AD # AD, DMS, AC
+  adi_lib: nothing
+  adi_rootname: "/SNL34m"
+  
+structuralParameters:
+  structuralModel: ROM #GX, TNB, ROM
+  nonlinear: false #TODO: propogate
+  ntelem: 10 #tower elements in each 
+  nbelem: 60 #blade elements in each 
+  ncelem: 10 #central cable elements in each if turbineType is ARCUS
+  nselem: 5 #strut elements in each if turbineType has struts
+  NuMad_geom_xlscsv_file_twr: /data/NuMAD_34m_TowerGeom.csv
+  NuMad_mat_xlscsv_file_twr: /data/NuMAD_34m_TowerMaterials.csv
+  NuMad_geom_xlscsv_file_bld: /data/NuMAD_SNL34mGeomBlades_AD.csv
+  NuMad_mat_xlscsv_file_bld: /data/NuMAD_SNL34mMaterials.csv
+  NuMad_geom_xlscsv_file_strut: /data/NuMAD_SNL34mGeomStruts.csv
+  NuMad_mat_xlscsv_file_strut: /data/NuMAD_SNL34mMaterials.csv
