+++ examples/helical_riverine/GeomStruts.csv
@@ -1,0 +1,5 @@
+SNL_34m_Aluminum,,,,# Shear webs ->,11,# of stacks ->,6,Stack ID ->,1,2,3,4,5,6,Segment ID ->,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,DP Types,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,# of segments ->,24,Material ID ->,1,2,3,4,5,6,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,first entry in the CSV list is the exterior blade surface stack,,,,,,,,,,,,,,,,,,,,,,,,SW1,,SW2,,SW3,,SW4,,SW5,,SW6,,SW7,,SW8,,SW9,,SW10,,SW11,
+Sta.#,Blade span (m),Airfoil Shape File,TE Type,Twist (deg),Chord,xoffset,aero center,Section Label,Base Skin and Rib,Mid Skin,Zero,Rib Mid,Rib Center,Center Skin,-blank-,HPBay12,HPBay11,HPBay10,HPBay9,HPBay8,HPBay7,HPBay6,HPBay5,HPBay4,HPBay3,HPBay2,HPBay1,LPBay1,LPBay2,LPBay3,LPBay4,LPBay5,LPBay6,LPBay7,LPBay8,LPBay9,LPBay10,LPBay11,LPBay12,,,,,,,,,,,,,,,,,,,,,,,,,,HPBay12,HPBay11,HPBay10,HPBay9,HPBay8,HPBay7,HPBay6,HPBay5,HPBay4,HPBay3,HPBay2,HPBay1,LPBay1,LPBay2,LPBay3,LPBay4,LPBay5,LPBay6,LPBay7,LPBay8,LPBay9,LPBay10,LPBay11,LPBay12,stacks,dps,stacks,,,,,,,,,,,,,,,,,,,
+1,0,NACA_0021,sharp,0,1.219,0.3,0.25,R,1,0,0,0,0,0,-1,-0.9167,-0.8333,-0.75,-0.6667,-0.5833,-0.5,-0.4167,-0.3333,-0.25,-0.1667,-0.0833,0,0.0833,0.1667,0.25,0.3333,0.4167,0.5,0.5833,0.6667,0.75,0.8333,0.9167,1,,,,,,,,,,,,,,,,,,,,,,,,,,"1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","13,11,11,13","1,","14,10,10,14","1,","15,9,9,15","1,","16,8,8,16","1,","17,7,7,17","1,","18,6,6,18","1,","19,5,5,19","1,","20,4,4,20","1,","21,3,3,21","1,","22,2,2,22","1,","23,1,1,23"
+2,1,NACA_0021,sharp,0,1.219,0.3,0.25,R,1,0,0,0,0,0,-1,-0.9167,-0.8333,-0.75,-0.6667,-0.5833,-0.5,-0.4167,-0.3333,-0.25,-0.1667,-0.0833,0,0.0833,0.1667,0.25,0.3333,0.4167,0.5,0.5833,0.6667,0.75,0.8333,0.9167,1,,,,,,,,,,,,,,,,,,,,,,,,,,"1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","1,","13,11,11,13","1,","14,10,10,14","1,","15,9,9,15","1,","16,8,8,16","1,","17,7,7,17","1,","18,6,6,18","1,","19,5,5,19","1,","20,4,4,20","1,","21,3,3,21","1,","22,2,2,22","1,","23,1,1,23"
