+++ examples/helical_riverine/NuMAD_SNL34mMaterials.csv
@@ -1,0 +1,9 @@
+,,,,Ex,Ey,Ez,Gxy,Gyz,Gxz,prxy,pryz,prxz,,,,
+,,Type,Layer Thickness,E11,E22,,G12,,,nu12,,,density,UTSx,UCSx,UTSy,UCSy,UTSz,UCSz,Cost,Reference,,
+Material ID,,,[mm],[MPa],[MPa],[MPa],[MPa],[MPa],[MPa],[-],[-],[-],[kg/m3],[MPa],[MPa],[MPa],[MPa],[MPa],[MPa],[$/kg],,,
+1,Aluminum_6063_T6_R,isotropic,8.087738693,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,Base Skin and Web
+2,Aluminum_6063_T6_T&C,isotropic,4.619000542,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,Mid Skin
+3,Aluminum_6063_T6_zero,isotropic,0,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,"For numerical ease, include webs etc everywhere, just set thickness to zero"
+4,Aluminum_6063_T6_T&C,isotropic,9.003307364,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,Rib Mid
+5,Aluminum_6063_T6_T&C,isotropic,9.563734009,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,Rib Center
+6,Aluminum_6063_T6_T&C,isotropic,4.806055859,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,100,100,100,100,3.00,Center Skin
