+++ examples/helical_riverine/TowerGeom.csv
@@ -1,0 +1,54 @@
+SNL_34m_Glass_BPE,,,,# Shear webs ->,0,# of stacks ->,2,Stack ID ->,1,2,Segment ID ->,1,10,DP Types,1,2,1,2,,,,,,,,,,,,,,,,,,,,
+,,,,,,# of segments ->,2,Material ID ->,8,8,,,,,,,first entry in the CSV list is the exterior blade surface stack,,,,,,,,,,,,,,,,,,,,,
+Sta.#,Blade span (m),Airfoil Shape File,TE Type,Twist (deg),Chord,xoffset,aero center,Section Label,Uni,Skin,-blank-,HP,LP,,,,HP,LP,,,,,,,,,,,,,,,,,,,,
+1,0,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+2,4.1789,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+3,8.3578,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+4,12.5367,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+5,16.7156,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+6,20.8945,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+7,25.0734,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+8,29.2523,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+9,33.4312,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+10,37.6101,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+11,41.789,circular,round,0,2,0.5,0.25,,24.000,1.000,-1.0000,0.0000,1.0000,,,,"2,1,2","2,1,2",,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
