+++ examples/helical_riverine/TowerMaterials.csv
@@ -1,0 +1,11 @@
+,,,,Ex,Ey,Ez,Gxy,Gyz,Gxz,prxy,pryz,prxz,,,,
+,,Type,Layer Thickness,E11,E22,,G12,,,nu12,,,density,UTS,UCS,Reference
+Material ID,,,[mm],[MPa],[MPa],[MPa],[MPa],[MPa],[MPa],[-],[-],[-],[kg/m3],[MPa],[MPa],
+1,ELT_5500,orthotropic,1.3,47835,18197,18197,2826,2826,2826,0.3,0.3,0.3,1950,875.634139,-592.949102,
+2,EBX_2400,orthotropic,0.61,17183,17183,17183,9202,9202,9202,0.3,0.3,0.3,1900,455.053962,-455.053962,
+3,ETLX_2400,orthotropic,0.66,20333,9305,9305,4756,4756,4756,0.3,0.3,0.3,1900,530.896289,-530.896289,
+4,Airex_C70_55,orthotropic,1,45,45,69,22,22,22,0.2,0.2,0.2,59,100,100,
+5,EBX_2400_x10,orthotropic,0.13,478353,181967,181967,28260,28260,28260,0.3,0.3,0.3,19000,4550.53962,-4550.53962,
+6,ETLX_2400_x10,orthotropic,0.06,171825,171825,171825,92019,92019,92019,0.3,0.3,0.3,19000,5308.96289,-5308.96289,
+7,Airex_C70_55_x10,orthotropic,0.07,203335,93051,93051,47560,47560,47560,0.2,0.2,0.2,590,0,0,
+8,Aluminum_6063_T6_T&C,isotropic,0.5,6.89E+04,6.89E+04,6.89E+04,2.58E+04,2.58E+04,2.58E+04,0.33,0.33,0.33,2700,241,434,
