+++ examples/helical_riverine/airfoils/NACA_0018.csv
@@ -1,0 +1,37 @@
+1.000001,0.0
+1.0000,0.00189
+0.9500,0.01210
+0.9000,0.02172
+0.8000,0.03935
+0.7000,0.05496
+0.6000,0.06845
+0.5000,0.07941
+0.4000,0.08705
+0.3000,0.09003
+0.2500,0.08912
+0.2000,0.08606
+0.1500,0.08018
+0.1000,0.07024
+0.0750,0.06300
+0.0500,0.05332
+0.0250,0.03922
+0.0125,0.02841
+0.0000,0.00000
+0.0125,-0.02841
+0.0250,-0.03922
+0.0500,-0.05332
+0.0750,-0.06300
+0.1000,-0.07024
+0.1500,-0.08018
+0.2000,-0.08606
+0.2500,-0.08912
+0.3000,-0.09003
+0.4000,-0.08705
+0.5000,-0.07941
+0.6000,-0.06845
+0.7000,-0.05496
+0.8000,-0.03935
+0.9000,-0.02172
+0.9500,-0.01210
+1.0000,-0.00189
+1.000001,0.0
