+++ examples/helical_riverine/airfoils/NACA_0021.csv
@@ -1,0 +1,37 @@
+100.0001,0
+100,0.221
+95,1.412
+90,2.534
+80,4.591
+70,6.412
+60,7.986
+50,9.265
+40,10.156
+30,10.504
+25,10.397
+20,10.04
+15,9.354
+10,8.195
+7.5,7.35
+5,6.221
+2.5,4.576
+1.25,3.315
+0,0
+1.25,-3.315
+2.5,-4.576
+5,-6.221
+7.5,-7.35
+10,-8.195
+15,-9.354
+20,-10.04
+25,-10.397
+30,-10.504
+40,-10.156
+50,-9.265
+60,-7.986
+70,-6.412
+80,-4.591
+90,-2.534
+95,-1.412
+100,-0.221
+100.0001,0
