+++ examples/helical_riverine/airfoils/circular.csv
@@ -1,0 +1,100 @@
+1,0
+0.999013364,-0.03139526
+0.996057351,-0.062666617
+0.991143625,-0.093690657
+0.984291581,-0.124344944
+0.975528258,-0.154508497
+0.964888243,-0.184062276
+0.952413526,-0.212889646
+0.93815334,-0.240876837
+0.922163963,-0.267913397
+0.904508497,-0.293892626
+0.885256621,-0.318711995
+0.864484314,-0.342273553
+0.842273553,-0.364484314
+0.818711995,-0.385256621
+0.793892626,-0.404508497
+0.767913397,-0.422163963
+0.740876837,-0.43815334
+0.712889646,-0.452413526
+0.684062276,-0.464888243
+0.654508497,-0.475528258
+0.624344944,-0.484291581
+0.593690657,-0.491143625
+0.562666617,-0.496057351
+0.53139526,-0.499013364
+0.5,-0.5
+0.46860474,-0.499013364
+0.437333383,-0.496057351
+0.406309343,-0.491143625
+0.375655056,-0.484291581
+0.345491503,-0.475528258
+0.315937724,-0.464888243
+0.287110354,-0.452413526
+0.259123163,-0.43815334
+0.232086603,-0.422163963
+0.206107374,-0.404508497
+0.181288005,-0.385256621
+0.157726447,-0.364484314
+0.135515686,-0.342273553
+0.114743379,-0.318711995
+0.095491503,-0.293892626
+0.077836037,-0.267913397
+0.06184666,-0.240876837
+0.047586474,-0.212889646
+0.035111757,-0.184062276
+0.024471742,-0.154508497
+0.015708419,-0.124344944
+0.008856375,-0.093690657
+0.003942649,-0.062666617
+0.000986636,-0.03139526
+0.0,0.0
+0.000986636,0.03139526
+0.003942649,0.062666617
+0.008856375,0.093690657
+0.015708419,0.124344944
+0.024471742,0.154508497
+0.035111757,0.184062276
+0.047586474,0.212889646
+0.06184666,0.240876837
+0.077836037,0.267913397
+0.095491503,0.293892626
+0.114743379,0.318711995
+0.135515686,0.342273553
+0.157726447,0.364484314
+0.181288005,0.385256621
+0.206107374,0.404508497
+0.232086603,0.422163963
+0.259123163,0.43815334
+0.287110354,0.452413526
+0.315937724,0.464888243
+0.345491503,0.475528258
+0.375655056,0.484291581
+0.406309343,0.491143625
+0.437333383,0.496057351
+0.46860474,0.499013364
+0.5,0.5
+0.53139526,0.499013364
+0.562666617,0.496057351
+0.593690657,0.491143625
+0.624344944,0.484291581
+0.654508497,0.475528258
+0.684062276,0.464888243
+0.712889646,0.452413526
+0.740876837,0.43815334
+0.767913397,0.422163963
+0.793892626,0.404508497
+0.818711995,0.385256621
+0.842273553,0.364484314
+0.864484314,0.342273553
+0.885256621,0.318711995
+0.904508497,0.293892626
+0.922163963,0.267913397
+0.93815334,0.240876837
+0.952413526,0.212889646
+0.964888243,0.184062276
+0.975528258,0.154508497
+0.984291581,0.124344944
+0.991143625,0.093690657
+0.996057351,0.062666617
+0.999013364,0.03139526
